# openapi.yaml
openapi: 3.1.0
info:
  title: FinSight AI API
  description: "Financial Intelligence & Decision Platform \u2014 Turn financial documents\
    \ into business decisions. Supports Gemini, OpenAI, and Ollama backends."
  version: 0.2.0
paths:
...
  /documents/:
    get:
      tags:
      - documents
      summary: List Documents
      operationId: list_documents_documents__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
      security:
      - HTTPBearer: []
...
  /query/:
    post:
      tags:
      - query
      summary: Query Agent
      operationId: query_agent_query__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /query/ratio:
    post:
      tags:
      - query
      summary: Calculate Ratio
      description: Safely calculate financial ratios like P/E, ROA, ROE, Margin, Debt-to-Equity.
      operationId: calculate_ratio_query_ratio_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinancialRatioRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialRatioResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /cases/:
    get:
      tags:
      - case-studies
      summary: Get Cases
      operationId: get_cases_cases__get
      security:
      - HTTPBearer: []
      parameters:
      - name: topic
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by topic (e.g. M&A, DCF Valuation)
          title: Topic
        description: Filter by topic (e.g. M&A, DCF Valuation)
      - name: difficulty
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by difficulty (Beginner, Intermediate, Advanced, Expert)
          title: Difficulty
        description: Filter by difficulty (Beginner, Intermediate, Advanced, Expert)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseStudy'
                title: Response Get Cases Cases  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CaseStudy:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        company_name:
          type: string
          title: Company Name
        ticker:
          type: string
          title: Ticker
          default: ''
        industry:
          type: string
          title: Industry
        topic:
          type: string
          enum:
          - M&A
          - DCF Valuation
          - Equity Research
          - Credit & Distress
          - Growth Investment
          - Corporate Finance
          title: Topic
        difficulty:
          type: string
          enum:
          - Beginner
          - Intermediate
          - Advanced
          - Expert
          title: Difficulty
        headline:
          type: string
          title: Headline
        description:
          type: string
          title: Description
        key_question:
          type: string
          title: Key Question
        context_dossier:
          type: string
          title: Context Dossier
        historical_financials:
          items:
            $ref: '#/components/schemas/FinancialHistorical'
          type: array
          title: Historical Financials
        current_metrics:
          additionalProperties: true
          type: object
          title: Current Metrics
        ground_truth_insights:
          items:
            type: string
          type: array
          title: Ground Truth Insights
        potential_risks:
          items:
            type: string
          type: array
          title: Potential Risks
        consensus_decision:
          type: string
          title: Consensus Decision
          default: BUY
        target_valuation_range:
          type: string
          title: Target Valuation Range
          default: $45B - $52B
      type: object
      required:
      - id
      - title
      - company_name
      - industry
      - topic
      - difficulty
      - headline
      - description
      - key_question
      - context_dossier
      title: CaseStudy
...
    DocumentListResponse:
      properties:
        sources:
          items:
            type: string
          type: array
          title: Sources
        total_chunks:
          type: integer
          title: Total Chunks
      type: object
      required:
      - sources
      - total_chunks
      title: DocumentListResponse
...
    FinancialHistorical:
      properties:
        year:
          type: string
          title: Year
        revenue:
          type: number
          title: Revenue
        cogs:
          type: number
          title: Cogs
        gross_profit:
          type: number
          title: Gross Profit
        operating_expenses:
          type: number
          title: Operating Expenses
        ebitda:
          type: number
          title: Ebitda
        net_income:
          type: number
          title: Net Income
        capex:
          type: number
          title: Capex
        fcf:
          type: number
          title: Fcf
      type: object
      required:
      - year
      - revenue
      - cogs
      - gross_profit
      - operating_expenses
      - ebitda
      - net_income
      - capex
      - fcf
      title: FinancialHistorical
    FinancialRatioRequest:
      properties:
        numerator:
          type: number
          title: Numerator
        denominator:
          type: number
          title: Denominator
        ratio_name:
          type: string
          title: Ratio Name
          default: custom_ratio
      type: object
      required:
      - numerator
      - denominator
      title: FinancialRatioRequest
    FinancialRatioResponse:
      properties:
        ratio_name:
          type: string
          title: Ratio Name
        value:
          anyOf:
          - type: number
          - type: 'null'
          title: Value
        numerator:
          type: number
          title: Numerator
        denominator:
          type: number
          title: Denominator
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - ratio_name
      - numerator
      - denominator
      title: FinancialRatioResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QueryRequest:
      properties:
        query:
          type: string
          maxLength: 2000
          minLength: 3
          title: Query
          description: Financial question to answer
        top_k:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: Top K
          default: 5
        stream:
          type: boolean
          title: Stream
          default: false
      type: object
      required:
      - query
      title: QueryRequest
    QueryResponse:
      properties:
        query:
          type: string
          title: Query
        answer:
          type: string
          title: Answer
        sources:
          items:
            type: string
          type: array
          title: Sources
        intent:
          type: string
          title: Intent
        agent_trace:
          items:
            type: string
          type: array
          title: Agent Trace
        tokens_estimated:
          anyOf:
          - type: integer
          - type: 'null'
          title: Tokens Estimated
      type: object
      required:
      - query
      - answer
      - sources
      - intent
      - agent_trace
      title: QueryResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer